ai_translator/pyproject.toml
[tool.poetry]
name = "ai-translator"
version = "0.1.0"
description = "RAG pipeline for translating scientific papers to accessible language"
authors = ["Antigravity"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-anthropic = "^0.2.0"
langchain-ollama = "^0.2.0"
langchain-chroma = "^0.1.0"
chromadb = "^0.5.0"
pypdf = "^5.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
mlflow = "^2.14.0"
textstat = "^0.7.4"
evaluate = "^0.4.2"
ragas = "^0.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
